[MSLITE][Develop] add cpu op: invert_permutation, size; add mode CropAndResize for resize op

+++ mindspore/lite/src/runtime/kernel/arm/base/resize_base.h
@@ -38,8 +38,8 @@ class ResizeBaseCPUKernel : public LiteKernel {
 
  protected:
   int method_ = 0;
-  int64_t new_height_ = 0;
-  int64_t new_width_ = 0;
+  int new_height_ = 0;
+  int new_width_ = 0;
   int coordinate_transform_mode_;
   bool preserve_aspect_ratio_ = false;
   bool const_shape_ = false;
